+++ sites/partners/src/pages/_app.tsx
@@ -14,10 +14,6 @@ import {
   RequireLogin,
 } from "@bloom-housing/shared-helpers"
 
-// TODO: Make these not-global
-import "ag-grid-community/dist/styles/ag-grid.css"
-import "ag-grid-community/dist/styles/ag-theme-alpine.css"
-
 import LinkComponent from "../components/core/LinkComponent"
 import { translations, overrideTranslations } from "../lib/translations"
