+++ sites/public/src/components/applications/ValidateAddress.tsx
@@ -1,6 +1,5 @@
 import { Address, MultiLineAddress, t } from "@bloom-housing/ui-components"
 import { Button } from "@bloom-housing/ui-seeds"
-import GeocodeService from "@mapbox/mapbox-sdk/services/geocoding"
 
 export interface FoundAddress {
   newAddress?: Address
@@ -13,17 +12,20 @@ export const findValidatedAddress = (
   setFoundAddress: React.Dispatch<React.SetStateAction<FoundAddress>>,
   setNewAddressSelected: React.Dispatch<React.SetStateAction<boolean>>
 ) => {
-  const geocodingClient = GeocodeService({
-    accessToken: process.env.mapBoxToken || process.env.MAPBOX_TOKEN,
-  })
+  void import("@mapbox/mapbox-sdk/services/geocoding")
+    .then(({ default: GeocodeService }) => {
+      const geocodingClient = GeocodeService({
+        accessToken: process.env.mapBoxToken || process.env.MAPBOX_TOKEN,
+      })
 
-  geocodingClient
-    .forwardGeocode({
-      query: `${address.street}, ${address.city}, ${address.state} ${address.zipCode}, United States`,
-      limit: 1,
-      types: ["address"],
+      return geocodingClient
+        .forwardGeocode({
+          query: `${address.street}, ${address.city}, ${address.state} ${address.zipCode}, United States`,
+          limit: 1,
+          types: ["address"],
+        })
+        .send()
     })
-    .send()
     .then((response) => {
       const [street, city, region] = response.body.features[0].place_name.split(", ")
       const coordinates = response.body.features[0].geometry?.coordinates
