+++ sites/public/src/components/content-pages/PrivacySeeds.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useContext } from "react"
 import { t } from "@bloom-housing/ui-components"
-import Markdown from "markdown-to-jsx"
+import LazyMarkdown from "../core/LazyMarkdown"
 import { PageView, pushGtmEvent, AuthContext } from "@bloom-housing/shared-helpers"
 import { UserStatus } from "../../lib/constants"
 import Layout from "../../layouts/application"
@@ -22,7 +22,7 @@ const Privacy = () => {
   return (
     <Layout pageTitle={t("pageTitle.privacy")}>
       <PageHeaderLayout heading={t("pageTitle.privacy")} inverse>
-        <Markdown className={styles["markdown"]}>{pageContent.toString()}</Markdown>
+        <LazyMarkdown className={styles["markdown"]}>{pageContent.toString()}</LazyMarkdown>
       </PageHeaderLayout>
     </Layout>
   )
