+++ sites/public/src/components/core/LazyMap.tsx
@@ -1,0 +1,16 @@
+import React from "react"
+import dynamic from "next/dynamic"
+
+const Map = dynamic(() => import("@bloom-housing/shared-helpers/src/views/components/Map"), {
+  ssr: false,
+})
+
+type LazyMapProps = {
+  [key: string]: unknown
+}
+
+const LazyMap = (props: LazyMapProps) => {
+  return <Map {...props} />
+}
+
+export default LazyMap
