+++ sites/public/src/layouts/application.tsx
@@ -1,7 +1,6 @@
 import React, { useContext, useEffect, useState } from "react"
 import dayjs from "dayjs"
 import { NextRouter, useRouter } from "next/router"
-import Markdown from "markdown-to-jsx"
 import Head from "next/head"
 import HomeIcon from "@heroicons/react/24/solid/HomeIcon"
 import { Message, Toast, Icon } from "@bloom-housing/ui-seeds"
@@ -9,6 +8,7 @@ import { MenuLink, t, SiteHeader as UICSiteHeader } from "@bloom-housing/ui-comp
 import { AuthContext, MessageContext } from "@bloom-housing/shared-helpers"
 import { User } from "@bloom-housing/shared-helpers/src/types/backend-swagger"
 import { ToastProps } from "@bloom-housing/ui-seeds/src/blocks/Toast"
+import LazyMarkdown from "../components/core/LazyMarkdown"
 import { MetaTags } from "../components/shared/MetaTags"
 import CustomSiteFooter from "../components/shared/CustomSiteFooter"
 import { HeaderLink, SiteHeader } from "../patterns/SiteHeader"
@@ -92,14 +92,14 @@ const getSiteHeaderDeprecated = (
       {showMaintenanceMessage && (
         <div className={`${styles["site-banner-container"]} ${styles["site-banner-alert-bg"]}`}>
           <Message className={styles["site-alert-banner-content"]} variant={"alert"}>
-            <Markdown>{t("alert.maintenance")}</Markdown>
+            <LazyMarkdown>{t("alert.maintenance")}</LazyMarkdown>
           </Message>
         </div>
       )}
       {showGenericSiteMessage && (
         <div className={`${styles["site-banner-container"]} ${styles["site-banner-primary-bg"]}`}>
           <Message className={styles["site-alert-banner-content"]} variant={"primary"}>
-            <Markdown>{t("siteMessage.generic")}</Markdown>
+            <LazyMarkdown>{t("siteMessage.generic")}</LazyMarkdown>
           </Message>
         </div>
       )}
