+++ sites/public/src/pages/applications/review/terms.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect, useMemo, useState } from "react"
 import { useRouter } from "next/router"
 import { useForm } from "react-hook-form"
-import Markdown from "markdown-to-jsx"
+import LazyMarkdown from "../../../components/core/LazyMarkdown"
 import { t, FieldGroup, Form, AlertBox } from "@bloom-housing/ui-components"
 import { CardSection } from "@bloom-housing/ui-seeds/src/blocks/Card"
 import {
@@ -180,17 +180,17 @@ const ApplicationTerms = () => {
             <div className="markdown">
               {listing?.applicationDueDate && (
                 <>
-                  <Markdown options={{ disableParsingRawHTML: true }}>
+                  <LazyMarkdown options={{ disableParsingRawHTML: true }}>
                     {t("application.review.terms.textSubmissionDate", {
                       applicationDueDate: applicationDueDate,
                     })}
-                  </Markdown>
+                  </LazyMarkdown>
                   <br />
                   <br />
                 </>
               )}
 
-              <Markdown
+              <LazyMarkdown
                 options={{
                   disableParsingRawHTML: true,
                   overrides: {
@@ -205,7 +205,7 @@ const ApplicationTerms = () => {
                 }}
               >
                 {content.text}
-              </Markdown>
+              </LazyMarkdown>
 
               <div className="mt-6">
                 <FieldGroup
