+++ sites/public/src/pages/applications/start/what-to-expect.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useContext, useMemo } from "react"
 import { useRouter } from "next/router"
 import { useForm } from "react-hook-form"
-import Markdown from "markdown-to-jsx"
+import LazyMarkdown from "../../../components/core/LazyMarkdown"
 import { ReviewOrderTypeEnum } from "@bloom-housing/shared-helpers/src/types/backend-swagger"
 import { t, Form } from "@bloom-housing/ui-components"
 import { OnClientSide, PageView, pushGtmEvent, AuthContext } from "@bloom-housing/shared-helpers"
@@ -93,7 +93,7 @@ const ApplicationWhatToExpect = () => {
       >
         <CardSection>
           <div className="markdown">
-            <Markdown
+            <LazyMarkdown
               options={{
                 disableParsingRawHTML: true,
                 overrides: {
@@ -108,9 +108,9 @@ const ApplicationWhatToExpect = () => {
               }}
             >
               {content.steps}
-            </Markdown>
+            </LazyMarkdown>
 
-            <Markdown
+            <LazyMarkdown
               options={{
                 disableParsingRawHTML: true,
                 overrides: {
@@ -125,7 +125,7 @@ const ApplicationWhatToExpect = () => {
               }}
             >
               {content.finePrint}
-            </Markdown>
+            </LazyMarkdown>
           </div>
         </CardSection>
         <CardSection className={styles["application-form-action-footer"]}>
