+++ sites/public/src/static_content/generic_faq_content.tsx
@@ -1,4 +1,4 @@
-import Markdown from "markdown-to-jsx"
+import LazyMarkdown from "../components/core/LazyMarkdown"
 import { Link } from "@bloom-housing/ui-seeds"
 import { t } from "@bloom-housing/ui-components"
 import { FaqCategory, FaqContent } from "../patterns/FrequentlyAskedQuestions"
@@ -10,9 +10,9 @@ export const getGenericFaqContent = (): FaqContent => {
       {
         question: t("faq.howDoesThisWork"),
         answer: (
-          <Markdown>{`${t("content.genericParagraph")} ${t("content.genericParagraph")} ${t(
+          <LazyMarkdown>{`${t("content.genericParagraph")} ${t("content.genericParagraph")} ${t(
             "content.genericParagraph"
-          )}`}</Markdown>
+          )}`}</LazyMarkdown>
         ),
       },
       {
